WaifuCore/requirements.txt
# requirements.txt

# Core AI & ML
openai
faster-whisper
torch
torchaudio
sentence-transformers # For memory embeddings
chromadb            # Vector database for long-term memory

# Web Framework & Utilities
httpx             # For async API calls
PyYAML

# numpy>=1.26.0  # Temporarily commented out due to Python 3.13 compatibility issues

# Audio
soundfile
pyaudio

kokoro>=0.9.2

gradio==4.31.5
fastapi==0.110.0
pydantic==2.6.4

TTS 